refactor(api.www.yantene.net): make LocalFile's sha1 static
