+++ sites/api.www.yantene.net/src/domain/aggregates/notes/value-objects/sha1.value-object.ts
@@ -1,3 +1,5 @@
+import { createHash } from "crypto";
+import { createReadStream } from "fs";
 import { ValueObjectInterface } from "../../../../common/interfaces/value-object.interface";
 
 export class Sha1 implements ValueObjectInterface {
@@ -9,6 +11,33 @@ export class Sha1 implements ValueObjectInterface {
 
   get value(): Buffer {
     return this.#value;
+  }
+
+  /**
+   * Calculate SHA-1 hash of the file.
+   *
+   * @param filePath - Path to the file
+   * @returns SHA-1 hash of the file
+   */
+  static async calculateFromFile(filePath: string): Promise<Sha1> {
+    const hash = createHash("sha1");
+    const stream = createReadStream(filePath);
+
+    return new Promise<Sha1>((resolve, reject) => {
+      stream.on("error", (error) => {
+        reject(error);
+      });
+
+      stream.on("data", (chunk) => {
+        hash.update(chunk);
+      });
+
+      stream.on("end", () => {
+        stream.close();
+
+        resolve(new Sha1(hash.digest()));
+      });
+    });
   }
 
   toJSON(): string {
